auth/supabase.py
from supabase import create_client, Client
import jwt
from fastapi import HTTPException, Security
from fastapi.security import HTTPBearer, HTTPAuthorizationCredentials
from config import settings

_supabase_client: Client = None

def get_supabase() -> Client:
    global _supabase_client
    if _supabase_client is None and settings.supabase_url and settings.supabase_service_role_key:
        _supabase_client = create_client(settings.supabase_url, settings.supabase_service_role_key)
    return _supabase_client


security = HTTPBearer()


def verify_token(credentials: HTTPAuthorizationCredentials = Security(security)) -> dict:
    """Decode and verify Supabase JWT. Returns the user payload."""
    if not settings.supabase_jwt_secret:
        raise HTTPException(status_code=500, detail="Auth not configured")
    try:
        payload = jwt.decode(
            credentials.credentials,
            settings.supabase_jwt_secret,
            algorithms=["HS256"],
            audience="authenticated",
        )
        return payload
    except jwt.ExpiredSignatureError:
        raise HTTPException(status_code=401, detail="Token expired")
    except jwt.InvalidTokenError:
        raise HTTPException(status_code=401, detail="Invalid token")


def require_admin(payload: dict) -> dict:
    """Check that the user has admin role in app_metadata."""
    app_meta = payload.get("app_metadata", {})
    if app_meta.get("role") != "admin":
        raise HTTPException(status_code=403, detail="Admin access required")
    return payload
