requirements.txt
fastapi>=0.111
uvicorn[standard]>=0.30
motor>=3.4
pymongo>=4.7
supabase>=2.4
PyJWT>=2.8
anthropic>=0.40
httpx>=0.27
python-dotenv>=1.0
pydantic>=2.7
pydantic-settings>=2.3
